+++ src/dsp_tools/commands/project/models/project_definition.py
@@ -1,12 +1,22 @@
+from __future__ import annotations
+
 from dataclasses import dataclass
+from typing import Any
 
 
 @dataclass
-class ProjectDefinition:
+class Project:
+    metadata: Metadata
+    ontologies: list[dict[str, Any]]
+    lists: list[dict[str, Any]] | None
+    groups: list[dict[str, str]] | None
+    users: list[dict[str, str]] | None
+
+
+@dataclass
+class Metadata:
     shortcode: str
     shortname: str
     longname: str
     keywords: list[str] | None = None
     descriptions: dict[str, str] | None = None
-    groups: list[dict[str, str]] | None = None
-    users: list[dict[str, str]] | None = None
